analytics/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0.post0
nats-py>=2.6